Add files via upload
I used the images from U.S. National Library of the Medicine  for the malaria disease database, it was a total mixed of 27558 images. For the purpose of the image processing before I fit my convolutional neural networks to them, I separated the images in two sub folders of the training set with 22000 images which was itself divided into the training infected and training uninfected cells. the rests of the pictures 2780 images which comprised 20% of the total images were grouped into the test sets for validation.
I used the keras to build my convolutional neural networks, I used two hidden layers of the network initially before adding one more layer for convolution in order to enhance the accuracy of my model. During 25 epoch I achieved the accuracy 0f 97.8 percent.